\name{Defunciones}
\alias{Defunciones}
\encoding{UTF-8}
\title{
Evaluación de los códigos de registros de defunción
}
\description{
Interfaz web para facilitar la evaluación de los códigos de defunción facilitando así el orden de los procesos.
}
\usage{
Defunciones()
}
\arguments{
 Función sin argumentos la entrada de datos es dentro de la interfaz
}
\details{
Mas detalles acerca de su uso se encuentran dentro de la interfaz web, ver en la pestaña de ayuda.
}
\value{
Las salidas están definidas por cada etapa.
}
\references{
Registros administrativos
}
\author{
José D. Loera
}

\examples{
## Función para evaluar códigos de  defunción.
if (interactive()) {
  Defunciones()
}
}
\keyword{ Registros }
