\name{SofiWebsite}
\alias{SofiWebsite}
\alias{Estadistica1}
\alias{Estadistica2}
\alias{Estadistica3}
\alias{Estadistica4}
\alias{Estadistica5}
\alias{Estadistica6}
\alias{Estadistica7}

\encoding{UTF-8}
\title{
Funciones de \pkg{Sofi} creadas para ser usadas en el lanzador Paper para más información visitar http://www.sofi.uno/.
}
\description{
Para saber más sobre cada función lanzar dicha función para mas detalles.
}

\author{Jose D. Loera \email{loerasg@gmail.com}}

\seealso{
 \code{\link{Sofi}}
}

\keyword{misc}
