\name{Estad}
\alias{Estad}
\encoding{UTF-8}
\title{
Para ejecutar diferentes niveles de estudio en estadística.
}
\description{
Es la función principal para ejecutar los diversos niveles de estadística, por lo pronto solo dos de prueba.
}
\usage{
Estad(Nivel)
}
\arguments{
  \item{Nivel}{
Para los parámetros
1 = Distribuciones diversas
2 = Graficas de datos
}
}
\details{
Para más detalles ver los ejemplos que se describirán abajo
}
\value{
No se tienen valores de salida
}
\references{
Varios de internet
}
\author{
José D. Loera
}

\examples{
##Estas es una función que lanza diferentes interfaces.

#Para ver distribuciones
#Estad(1)

#Para ver graficas de datos
#Estad(2)
}
\keyword{ Graficas }
\keyword{ Estadística }
