\name{Estad}
\alias{Estad}
\encoding{UTF-8}
\title{
Para ejecutar diferentes niveles de estudio en estadística.
}
\description{
Es la función principal para ejecutar los diversos niveles de estadística, por lo pronto solo dos de prueba.
}
\usage{
Estad(Nivel)
}
\arguments{
  \item{Nivel}{
  Para los parámetros
  \tabular{ll}{
1 = \tab Calculadora de Distribución\cr
2 = \tab Distribuciones de probabilidad\cr
3 = \tab Menú de Sofi\cr
4 = \tab Gráficas de datos\cr
}
}
}
\details{
Para más detalles ver los ejemplos que se describirán abajo
}
\value{
No se tienen valores de salida
}
\references{
Varios de internet
}
\author{
José D. Loera
}

\examples{
## Esta es una función que lanza diferentes interfaces.

# Para ver una Calculadora de Distribución
# Estad(1)

# Para ver Distribuciones de probabilidad
# Estad(2)

# Menú de Sofi
# Estad(3)

# Graficas de datos
# Estad(4)
}
\keyword{ Graficas }
\keyword{ Estadística }
